Refactor internals into recorders and renderers
A recorder is given a frame object and should record the time spent in
that frame. At the end of the run the recorder supplies the data
structure for the renderer to render.

The renderer turns the data structure (a tree of Frame objects) into
a rendered output (e.g. HTML, console text)