SpeedscopeRenderer: implement speedscope format
This commit replaces the SpeedscopeRenderer implementation copied from
JSONRenderer with an implementation of an "evented" speedscope JSON
profile output file writer.

The file format is documented in a few different ways, listed below:

* the speedscope GitHub repo wiki:
  https://github.com/jlfwong/speedscope/wiki/Importing-from-custom-sources

* the speedscope JSON schema:
  https://www.speedscope.app/file-format-schema.json

* a Typescript file documenting the schema
  https://github.com/jlfwong/speedscope/blob/master/src/lib/file-format-spec.ts

* a simple example
  https://github.com/jlfwong/speedscope/blob/100578c536a3afab39fb6803d28913d12eac29c5/sample/profiles/speedscope/0.0.1/simple.speedscope.json

The style of the implementation needs obvious work; this commit is
intended as a starting point for a pull request with the expectation
that upstream maintainer feedback will be needed to fix any style or
documentation issues.