SpeedscopeRenderer: fix inconsistency in endValue
This commit fixes an apparent inconsistency in the
profiles[0].endValue field of the Speedscope output from
SpeedscopeRenderer. In unit testing, the value of session.duration
(within a call to SpeedscopeRenderer.render) may not be equal to the
event time of the last event in the profiles[0].events field of the
Speedscope output. In fact, in the test_speedscope_output test within
test_profiler.py, the value of session duration was approximately
0.0047s (in local testing on my laptop), whereas the time value of the
last event generated by the profile in that test should be 0.75s +/-
0.3s.

To correct this inconsistency, the end value of profile is set equal
to the event time of the last event.